// comonl-comonlwcl/src/app/modules/comonlapi/model/anagraficaAzienda.ts
/**
 * ComonlWeb
 * API per ComonlWeb
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { ApiError } from './apiError';


export interface AnagraficaAzienda { 
    apiWarnings?: Array<ApiError>;
    id?: number;
    codiceFiscale?: string;
    partitaIva?: string;
    denominazione?: string;
    tipoStudio?: string;
    comune?: string;
    provincia?: string;
    tipologia?: string;
    idChiave?: string;
    dataAnnullamento?: Date;
}